howto: update content to work with kernel v4.7-rc1
Hi Mike,

I know you're on vacation but I'm sending you this patch in the hope you'll
have time to have a quick glance at it.  It simply modifies the howto.md
to work with other kernels than bkk16.  The information has been outdated
for a while now but I was waiting to start pushing the perf user space tools
patches to update everything.

The first set of the perf user space tools patches is ready to be published,
something I intend to do tomorrow.  For that to happen I need to apply this
patch.  It is not a big deal if you don't catch this email in time - I will
simply apply the patch for now and we can deal with any changes that need
to be done (if any) when you are back.

Thanks,
Mathieu

Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>