opencsd: Update memory accessor API on decode tree.
Memory accessors allow trace analyser to follow opcodes of program traced.
The C-API has a relatively simple _add_accessor-...() set of functions.

The decode tree allowed access to the accessor map object which collates
all the accessors by address and a couple of functions to add to the map.

This results in the decode tree API for memory accessors requiring some
knowledge of the memory mapper, plus this resulted in a lot of intermediate
code in the C-API wrapper to do all the mapping operations.

The decode tree now also has a simple _AddAccessor interface, all the
intermediate code now being inside the decode tree object.

This makes the C-API and decode tree API consistent and the C++ library
easier to used.

Signed-off-by: Mike Leach <mike.leach@linaro.org>