opencsd: Update to ensure that the C packet structure is used in C API.
The ETMv3 packet C++ object has the C packet stuct encapsulated rather than
as a base class. This resulted in the wrong pointer value being passed out to
the C API. This has been corrected using a base packet class interface to
explicitly get a pointer to the C struct declared for the C API interface.

Signed-off-by: Mike Leach <mike.leach@linaro.org>