opencsd: Custom External Decoder C-API (adds factory and wrappers)
Adds a custom factory API for registering named decoders with the library and
creating new instances of the decoder types.

Updates to the decoder manager and decoder instance wrapper C++ objects that
interface with the C-API

Added in LogMessage to TraceComponent base class to allow derived classes
to easily call logging interface without deref pointers etc.

Signed-off-by: Mike Leach <mike.leach@linaro.org>