opencsd: Allow FSYNCs to be used as special frames.
When collecting on device trace, allow 4 * FSYNC frames to be used as
padding or barrier frames to split data from trace runs.

Allows linux drivers to work correctly with tools such as perf that will
grab multiple buffers of data and concatenate them in the output without
the ability for the decoder to spot the breaks.

Updates frame indexing to account for skipped FSYNCs on resets.
Allows frame index to be forwarded downstream reset operation.

Signed-off-by: Mike Leach <mike.leach@linaro.org>