opencsd: Add return stack object to library
Add class to library to handle return stack feature in trace for
PTM and ETMv4.

Return stack allows return addresses to be saved on an internal
hardware stack, and reused if matching on an indirect branch.
This reduces the need for explicit address packets on indirect branches.

The return stack object maintains a stack of at least the size required
by the protocols, as a circular buffer as old, unused values fall off
the end. the object allows for push() and pop() operations, a check for
overflow and also a flag to allow the decoder to mark pop as pending.
This is where the decision to pop() may be dependent on a following
packet.

Compile time debug option allows test and debug of the stack operations.

Signed-off-by: Mike Leach <mike.leach@linaro.org>