opencsd: Creating user API include directory
Some of the header file need to be exposed to client programs so that
they can be compiled against them.  As such moving the header file needed
to compile the current perf tools to an "opencsd" directory under
include/ and adjust other files to reflect the change.

With this change user programs can now be compiled without having
to specify the -I option.  The other advantage is that the library
can be packaged for installation on a system.

Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>