build: Adding "install" target
That way header and library files can end up in the right place
on the file system, and perf tools compiled without having to
specify any environment variable.

Limit the header files copied to those necessary for installaion of C-API
interface to the library.

Install list maintained in the C-API makefile.

Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Signed-off-by: Mike Leach <mike.leach@linaro.org>