docs: HOWTO update for changes making openCSD installable in linux
OpenCSD can now be installed as part of the linux system libraries.

It has been decided that "libopencsd" would be more intuitive than
"libcstraced", hence the modification found in this patch.

Now that the header and object files have been installed as
a system library there is no need to use the CSTRACE_PATH
environment variable when compiling the perf tools and and set
LD_LIBRARY_PATH to execute the new 'perf' binary.

HOWTO.md changed to reflect all these changes.

Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>