opencsd: bugfix: crash when accessing decode memory image.
The caching mechanism use to speed up memory image access can cause
crash when first access is to a memory area where an image for the address
has not been supplied for that area. (would normally emit a _ADDR_NACC
element.)

Caching logic has been fixed to correct this issue.

Signed-off-by: Mike Leach <mike.leach@linaro.org>