build: Use $(MAKE) for recursive make calls.
Previously the decoder makefile called recursive
make by calling make directly.

According to the make documentation this should not
be used[1]. Instead $(MAKE) variable should be used.
One of the problems that this solved in my case was
that I was not able to build opencsd in parallel because
the jobservers did not recognize children make.

[1] https://www.gnu.org/software/make/manual/html_node/MAKE-Variable.html