build: Map dependencies correctly.
Previously the dependencies were incorectly
marked on phony targets instead of the targets.

For example libraries did not have dependencies
on objects, only the phony target had, so Make
was free to build the objects and the library
archiving at the same time. This wrong as the
archiving actually depends on object files being
available

This leads to race conditions and failures to
build when OpenCSD build in parallel.

It may lead to more nasty errors like old object
files mixed with newly generated ones leading to
obvious defects.