build: copy_libs now marks propper dependencies.
copy_libs was run assuming that previous targets produced
shared objects. The problem is that if for example we
are running in parallel, make builds the PHONY copy_libs
target before the actual libraries that copy_libs depends
on are available, thus aborting the build.

This commit maps the dependencies of copy_libs appropriately.