opencsd: Add API for callback mem accessor with trace ID parameter.
Clients using the memory accessor callback function may require the trace ID
of the trace source being decoded when the memory access is requested to
correctly associate the cpu with the memory and obtain the appropriate data.

Memory access API updated to add callback funciton signature with this extra
parameter, and underlying code adjusted to provide the additional parameter.

Signed-off-by: Mike Leach <mike.leach@linaro.org>