opencsd: etmv4: Fix exception trace issue where exception at branch target.
Where an exception occurs at the target address of a taken branch, this can
result in no code being executed at the branch target address.

In this case, the address supplied in the exception packet can also form the
address of the branch target, and contain an optional context element to
indicate a context change due to the branch.

The exception handling code for this case was incorrect. This has been fixed.
A context element is now correctly emitted prior to the exception element if
indicated in the exception packets.

Additionally a flag to indicate to clients that the preferred return address
is also a branch target address has been added to the generic output
packets. This could be implied from the packet flow, but is explicitly flagged
to make detection easier.

Signed-off-by: Mike Leach <mike.leach@linaro.org>