opencsd: test: docs: Update core name / profile matching.
Update the core name to architecture version and profile matching class.
Adds in generic profile names such as ARMv8-A which can now be used in snapshots where
the core name is not supported in the library routine.

Added documentation for the mapping class and a note to the test program usage docs to ensure
that the valid name strings are full documented where needed.

Signed-off-by: Mike Leach <mike.leach@linaro.org>