opencsd: etmv4: bugfix: Issue with context following address on A64 -> A32 transition
A decode problem will occur if:-
a) a transition from 64 bit arch to 32 bit arch occurs.
b) this transition is signalled by an address packet followed by a separate context packet
(not a combined address + context packet)
c) the first waypoint after the transition is an opcode in A32 that is not a waypoint in A64.

This patch ensures that the decoder is immune to the separation and ordering of target address and
context packets in this situation.

Signed-off-by: Mike Leach <mike.leach@linaro.org>