opencsd: etmv4: TraceInfo packet - ensure addressing zeroed.
TraceInfo packet zero out the stored addresses to 0 - this was not happening for all occurances of trace info packets after the initial trace sync.

This could result in a rare corner case if TInfo occurs before an address packet that output after the TInfo for a low address when the previous addresses had high bits set and the implementation chooses to use a shorted address bit packet than the full address width.

Signed-off-by: Mike Leach <mike.leach@linaro.org>