opencsd: Add marker generic element type.
Some trace packets can act as positional markers in the trace stream for packets that
are output later to allow the analyzer to associate the later packet with the correct position.

This adds a generic element type to handle these marker packets and put a generic packet
in the output stream for the same reason.

Signed-off-by: Mike Leach <mike.leach@linaro.org>