tests: Update linux test script.
The test script is updated to allow a parameter input to specify the path to
the library / test program to be used.

By default the behaviour remains the same using a path of ./bin/linux64/rel,
an input of 'use-installed' will remove any paths so that the installed
versions are used, any other input will be used as a path to the test
program + libraries.

Signed-off-by: Mike Leach <mike.leach@linaro.org>