opencsd: build: fixes for rpm based distros
The default install target is installing the .so files
with owner rw, and r for group and owner. This causes ldd
and various packaging utilities on rpm based distros to
misunderstand that this is a shared library. Further, recent
packaging guidelines on fedora discourage the use of static
libraries. To support this, lets allow the current 'make install'
behavior to be overridden with a couple env variables.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>