opencsd: etm4: Inconsistent error handling.
Error handling in etm4 decoder was an inconsistent mix of halt on
bad packet, and continue, reset and wait for next sync.

This has been fixed to the stated default of continue and sync
for all cases.

Decoder create flags modified to force halt on bad packet or
sequence separately from error classification of error / warning.

Signed-off-by: Mike Leach <mike.leach@linaro.org>