opencsd: Remove noisy printf
This printf is the only occurrence outside of the tests. When using
OpenCSD with perf in TUI mode, printfs can corrupt the UI and make
perf's own error messages hard to see. The proper way to print would be
pr_warning() or ui__warning() which aren't available here.

I don't see an easy way of making this work, and I don't think this
printf is that useful either, so remove it.

Signed-off-by: James Clark <james.clark@arm.com>