opencsd: etm4x: Update etm4x / ete decoder to flush context to client immediately.
Handling for speculative trace added caching of trace elements to be committed cancelled.
This results in elements being output in blocks on commit.

This introduced an issue where a context element could change a context, but not be output to the client,
before subsequent atoms were processed, and memory accesses to the client could occur without the new
context being seen. This could result in the incorrect memory being served from the client to the decoder,
resulting in decode issues.