opencsd: frame demux: fix handling of stream with HSYNC and FSYNC (github #50)
The handling for trace streams using HSYNC and FSYNC packets was not
correctly accounting for an odd number of HSYNC packets in a stream,
which could result in read past end of buffer.

Routine changed to ensure handling of 2 byte aligned end of buffer,
and FSYNC across a pair of input buffers.

Reported-by: yabinc <yabinc@google.com> (github issue #50)
Signed-off-by: Mike Leach <mike.leach@linaro.org>