opencsd: etmv4: TINFO - fix spec trace handling. (github #75)
The current decoder resets the return stack on seeing a periodic trace
info element, without adding that element to the speculative element s
tack.

This is a potential problem if elements are changed / cancelled due to
speculation around the trace info element

Furthermore, the spec prohibits push to the return stack for any elements
 occurring between the trace info element and the required address output
after the trace info element.

Alter the decoder to add trace info as an element on the P0 element stack,
and handle the trace info requirements during speculation resolution.

Signed-off-by: Mike Leach <mike.leach@linaro.org>